feat: Enhance TypeScript configurations and improve Prebid integration
- Updated tsconfig.json in common and extension packages to include prebidGlobal types.
- Added `prepareTimelineData` utility function to process auction events in the design system.
- Refactored Timeline component to utilize Bidder interface and improved type safety.
- Removed deprecated prebid.d.ts file and integrated necessary types into index.d.ts.
- Enhanced PrebidBidsPanel to utilize the new prepareTimelineData function for better data handling.
- Added additional properties to Bidder interface for more comprehensive data representation.
- Updated various components and stories to reflect changes in data structures and types.
