+++ .gitignore
@@ -36,3 +36,4 @@ tmp
 
 
 *storybook.log
+**/storybook-static
