+++ packages/design-system/src/components/index.ts
@@ -73,6 +73,7 @@ export { default as CardsPanel } from './landingPage/cardsPanel';
 export { default as FrameContent } from './frameContent';
 export { default as JsonView } from './jsonView/jsonView';
 export { default as Timeline } from './timeline';
+export { default as prepareTimelineData } from './timeline/utils/prepareTimelineData';
 export * from './timeline/types';
 export { default as MultiSelectDropDown } from './multiSelectDropDown';
 export { default as Dropdown } from './dropdown';
