+++ packages/extension/src/@types/index.d.ts
@@ -13,6 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/**
+ * External dependencies.
+ */
+import type { PrebidJsGlobal } from '@google-psat/common';
 declare module '*.svg' {
   import React = require('react');
   const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
@@ -37,3 +41,9 @@ export type RelatedWebsiteSetType = {
     [url: string]: string;
   };
 };
+
+declare global {
+  interface Window {
+    pbjs: PrebidJsGlobal;
+  }
+}
