+++ packages/extension/src/store/types.ts
@@ -16,12 +16,7 @@
 /**
  * External dependencies.
  */
-import type {
-  AdsAndBiddersType,
-  CookieData,
-  PrebidNoBidsType,
-  ReceivedBids,
-} from '@google-psat/common';
+import type { CookieData } from '@google-psat/common';
 
 export type PreferenceKeyValues =
   | 'columnSorting'
@@ -52,23 +47,4 @@ export type AllowedDomainObject = {
 
 export type AllowedDomainStorage = {
   allowList: AllowedDomainObject[];
-};
-
-export type PrebidWarningTypes = 'WARNING' | 'ERROR' | 'INFO';
-
-export type PrebidEvents = {
-  prebidExists: boolean | null;
-  adUnits: AdsAndBiddersType;
-  noBids: PrebidNoBidsType;
-  versionInfo: string;
-  receivedBids: ReceivedBids[];
-  errorEvents: {
-    type: PrebidWarningTypes;
-    message: string[];
-    time: string;
-  }[];
-  auctionEvents: { [auctionId: string]: any[] };
-  installedModules: string[];
-  config: PrebidConfig;
-  pbjsNamespace: string;
 };
